bugfix: Apply locking to itemCount() method
Prevents in-accurate counting of the items left in the queue.
As file locks are advisory, they need to be applied here to prevent a reader getting a partial count of the items left
in the queue whilst another thread is re-writing the file as part of the getItem() or addItem() methods.